requirements.txt
azure-functions
azure-identity
azure-monitor-query
openai
requests~=2.23.0